optimize home page data fetching and processing
