+++ apps/club/src/app/page.tsx
@@ -1,5 +1,3 @@
-import type { ReturnEvent } from "@forge/db/schemas/knight-hacks";
-
 import { env } from "~/env";
 import About from "./_components/landing/about";
 import CalendarPage from "./_components/landing/calendar";
@@ -10,21 +8,10 @@ import Sponsors from "./_components/landing/sponsors";
 import { api } from "./trpc/server";
 
 export default async function HomePage() {
-  const events = await api.event.getEvents.query();
-  const memberCount = await api.member.getMemberCount.query();
-
-  const eventMap = new Map<string, ReturnEvent[]>();
-
-  events.forEach((event) => {
-    const day = event.start_datetime.getDate();
-    const month = event.start_datetime.getMonth();
-    const year = event.start_datetime.getFullYear();
-    const dateString = `${day}-${month}-${year}`;
-    if (!eventMap.has(dateString)) {
-      eventMap.set(dateString, []);
-    }
-    eventMap.get(dateString)?.push(event);
-  });
+  const [events, memberCount] = await Promise.all([
+    api.event.getEvents.query(),
+    api.member.getMemberCount.query(),
+  ]);
 
   return (
     <div className="bg-[#0F172A]">
@@ -32,7 +19,7 @@ export default async function HomePage() {
       <About />
       <Impact />
       <Sponsors />
-      <CalendarPage events={eventMap} />
+      <CalendarPage events={events} />
       <Discover memberCount={memberCount} />
     </div>
   );
